# api/api-spec.yml
openapi: 3.1.0
servers:
  - url: http://localhost:3000
info:
  title: Inquiry Server
  description: The Inquiry Server API
  version: 0.0.1
  contact:
    name: Wim Spaargaren
    url: "https://github.com/inquiryproj/inquiry"
    email: "wim_spaargaren@live.nl"
...
components:
  schemas:
...
    Scenario:
      type: object
      required:
        - id
        - name
        - spec_type
        - spec
        - project_id
      properties:
        id:
          x-go-type: uuid.UUID
          x-go-name: ID
          x-go-type-import:
            path: github.com/google/uuid
        name: 
          type: string
        spec_type: 
          type: string
        spec: 
          type: string
        project_id:
          x-go-type: uuid.UUID
          x-go-name: ProjectID
          x-go-type-import:
            path: github.com/google/uuid
...
    ErrMsg:
      type: object
      required:
        - message
      properties:
        message:
          type: string
paths:
...
  "/v1/projects/{id}/scenarios":
    post:
      description: Creates a scenario
      operationId: createScenario
      tags:
        - scenarios
        - create
      parameters:
        - in: path
          name: id
          schema:
            type: string
            x-go-type: uuid.UUID
            x-go-name: ID
            x-go-type-import:
              path: github.com/google/uuid
          required: true
      requestBody:
        required: true
        content: 
          application/json:
            schema: 
              $ref: "#/components/schemas/Scenario"
      responses: 
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
          description: The scenario was successfully created.
        default:
          description: Unable to create scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrMsg"
...